refactor: 클래스명 및 디렉터리 변경
